# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.createUserRequest:
    properties:
      address:
        type: string
      avatar:
        type: string
      email:
        type: string
      enable:
        type: boolean
      gender:
        type: integer
      nickName:
        type: string
      password:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  controllers.resetPasswordRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
...
  controllers.updateProfileRequest:
    properties:
      address:
        type: string
      avatar:
        type: string
      email:
        type: string
      gender:
        type: integer
      nickName:
        type: string
    type: object
...
  controllers.updateUserRequest:
    properties:
      address:
        type: string
      avatar:
        type: string
      email:
        type: string
      enable:
        type: boolean
      gender:
        type: integer
      nickName:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      username:
        type: string
    type: object
...
  models.PageData:
    properties:
      pageData: {}
      total:
        type: integer
    type: object
...
  models.Profile:
    properties:
      address:
        type: string
      avatar:
        type: string
      email:
        type: string
      gender:
        type: integer
      id:
        type: integer
      nickName:
        type: string
      userId:
        type: integer
    type: object
...
  models.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      originUrl:
        type: string
    type: object
  models.Role:
    properties:
      code:
        type: string
      enable:
        type: boolean
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.User:
    properties:
      createTime:
        type: string
      enable:
        type: boolean
      id:
        type: integer
      profile:
        $ref: '#/definitions/models.Profile'
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updateTime:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: kite-admin 后台管理系统 API 文档
  title: Kite Admin API
  version: "1.0"
paths:
...
  /user:
    get:
      description: 分页查询用户列表，支持按用户名搜索
      parameters:
      - description: 用户名（模糊搜索）
        in: query
        name: username
        type: string
      - default: 1
        description: 页码
        in: query
        name: pageNo
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PageData'
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 创建新用户，可同时设置资料和分配角色
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.createUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - 用户管理
  /user/{id}:
    delete:
      description: 根据ID删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    patch:
      consumes:
      - application/json
      description: 更新用户基本信息、资料和角色
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 更新用户
      tags:
      - 用户管理
  /user/detail:
    get:
      description: 获取当前登录用户的详细信息，包括资料和角色
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 获取当前用户详情
      tags:
      - 用户管理
  /user/password/reset/{id}:
    patch:
      consumes:
      - application/json
      description: 管理员重置指定用户的密码
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 新密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.resetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 重置密码
      tags:
      - 用户管理
  /user/profile/{id}:
    patch:
      consumes:
      - application/json
      description: 用户更新自己的个人资料（昵称、头像、性别、地址、邮箱），仅允许更新本人
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 个人资料
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.updateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: 无权修改他人资料
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 更新个人资料
      tags:
      - 用户管理
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"